// frontend/src/lib/api-client/models/ResourceRepresentation.ts
export interface ResourceRepresentation {
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     */
    type?: string;
    /**
     * 
     * @type {ResourceOwnerRepresentation}
     * @memberof ResourceRepresentation
     */
    owner?: ResourceOwnerRepresentation;
    /**
     * 
     * @type {boolean}
     * @memberof ResourceRepresentation
     */
    ownerManagedAccess?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     */
    displayName?: string;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof ResourceRepresentation
     */
    attributes?: { [key: string]: Array<string>; };
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     */
    id?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof ResourceRepresentation
     */
    uris?: Set<string>;
    /**
     * 
     * @type {Set<ScopeRepresentation>}
     * @memberof ResourceRepresentation
     */
    scopes?: Set<ScopeRepresentation>;
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     */
    iconUri?: string;
    /**
     * 
     * @type {string}
     * @memberof ResourceRepresentation
     * @deprecated
     */
    uri?: string;
    /**
     * 
     * @type {Set<ScopeRepresentation>}
     * @memberof ResourceRepresentation
     */
    resourceScopes?: Set<ScopeRepresentation>;
}